requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.3.1
cron_descriptor==2.0.6
cycler==0.12.1
Django==4.2.28
django-celery-beat==2.8.1
django-extensions==4.1
django-filter==25.2
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
fastapi==0.115.11
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
hf-xet==1.1.0
hub-sdk==0.0.20
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.6.2
labelImg==1.8.6
lxml==5.4.0
Markdown==3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
mysqlclient==2.2.8
networkx==3.4.2
numpy==2.1.1
onnx==1.17.0
onnxruntime==1.21.1
onnxruntime-tools==1.7.0
onnxslim==0.1.51
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pillow_heif==0.22.0
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.0.0
psycopg2-binary==2.9.11
reportlab==4.2.5
py-cpuinfo==9.0.0
py3nvml==0.2.7
pybboxes==0.1.6
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.11.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
redis==7.2.0
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.63
sahi==0.11.23
scipy==1.15.2
seaborn==0.13.2
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
starlette==0.46.1
sympy==1.14.0
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.3
tzlocal==5.3.1
ultralytics==8.3.129
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.6.0
xmltodict==0.14.2
